Correcting the parsing

+++ srcs/color/color.c
@@ -25,9 +25,9 @@ rgba    atocolor(char *s)
         return (c);
     if (string_array_size(colors) < 3 || string_array_size(colors) > 4)
         return (free_string_array(colors), c);
-    c.r = (float)(ft_atoi(colors[0]) / 255);
-    c.g = (float)(ft_atoi(colors[1]) / 255);
-    c.b = (float)(ft_atoi(colors[2]) / 255);
+    c.r = ((float)ft_atoi(colors[0]) / (float)255);
+    c.g = ((float)ft_atoi(colors[1]) / (float)255);
+    c.b = ((float)ft_atoi(colors[2]) / (float)255);
     if (string_array_size(colors) == 4)
         c.a = ft_atof(colors[3], NULL);
     free_string_array(colors);
